% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_mix_functions.R
\name{cells_to_genotypes}
\alias{cells_to_genotypes}
\title{this is a function that maps the genotype clusters to genotype names}
\usage{
cells_to_genotypes(SOC_locations, membership_mat)
}
\arguments{
\item{SOC_locations}{path to the .csv file giving channel names as the first column and file paths to the souporcell output in the second column}

\item{membership_mat}{output of membership_map()}
}
\value{
matrix of cells by genotypes
cells_to_genotypes()
}
\description{
this is a function that maps the genotype clusters to genotype names
}
