% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_mix_functions.R
\name{check_complete_graph}
\alias{check_complete_graph}
\title{this is a big function that checks if the genotype graph is a set of complete subgraphs}
\usage{
check_complete_graph(gr)
}
\arguments{
\item{gr}{an igraph object}
}
\value{
a message telling you if your graph is a collection of complete subgraphs
}
\description{
this is a big function that checks if the genotype graph is a set of complete subgraphs
}
\examples{
\donttest{
check_complete_graph()
}
}
