% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cluster_color_ramp}
\alias{cluster_color_ramp}
\title{cluster_color_ramp}
\usage{
cluster_color_ramp(
  cols = c("aquamarine", "orange", "blue", "grey", "red", "pink", "deepskyblue",
    "yellow", "coral", "darkseagreen1"),
  ...
)
}
\arguments{
\item{cols}{vector of colours.}
}
\description{
cluster_color_ramp
}
