% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_mix_functions.R
\name{construct_genotype_cluster_graph}
\alias{construct_genotype_cluster_graph}
\title{this is a big function that constructs a genotype cluster graph using a set of SOC directories and an experimental design}
\usage{
construct_genotype_cluster_graph(experimental_design, SOC_locations)
}
\arguments{
\item{experimental_design}{an experimental design matrix rownames should be microfluidics channels, colnames should be genotypes}
}
\value{
a list $graph_membership gives you cluster memberships $graph_plot gives a force directed embedding of the graph $membership_plot gives a heatmap of memberships $membership_matrix gives a matrix of channel memberships
construct_genotype_cluster_graph()
}
\description{
this is a big function that constructs a genotype cluster graph using a set of SOC directories and an experimental design
}
