% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_mix_functions.R
\name{construct_genotype_cluster_graph_unsupervised}
\alias{construct_genotype_cluster_graph_unsupervised}
\title{Function that clusters genotypes into donors in an unsupervised manner (i.e. does not require the experimental design). This is useful if there is some doubt about the experimental design, or if one is not available.}
\usage{
construct_genotype_cluster_graph_unsupervised(
  file_locations,
  ncounts = 10,
  use_VAF = TRUE
)
}
\arguments{
\item{file_locations}{the file locations}

\item{ncounts}{numeric - the number of counts supporting each variant}

\item{use_VAF}{if TRUE calculates genotype to genotype similarity on the basis of mean squared error between variant allele frequencies, if FALSE calculates relatedness (Pedersen et al. 2020) using a reimplementation of the Somalier approach.}
}
\value{
a list $graph_membership gives you cluster memberships $graph_plot gives a force directed embedding of the graph $membership_plot gives a heatmap of memberships $membership_matrix gives a matrix of channel memberships, $histogram gives a histogram of relatedness values used to cluster links into donors/generate a graph.
}
\description{
Function that clusters genotypes into donors in an unsupervised manner (i.e. does not require the experimental design). This is useful if there is some doubt about the experimental design, or if one is not available.
}
\examples{
\donttest{
construct_genotype_cluster_graph()
}
}
