% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_mix_functions.R
\name{membership_map}
\alias{membership_map}
\title{this is a function that maps channel membership to genotype ID}
\usage{
membership_map(experimental_design, graph_output)
}
\arguments{
\item{experimental_design}{design matrix rownames should be microfluidics channels, colnames should be genotypes}

\item{graph_output}{the output of construct_genotype_cluster_graph given as x$genotype_cluster_graph}
}
\value{
a matrix of channel memberships for each cluster
}
\description{
this is a function that maps channel membership to genotype ID
}
\examples{
\donttest{
membership_map()
}
}
