% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_mix_functions.R
\name{plot_experimental_design}
\alias{plot_experimental_design}
\title{plot_experimental_design}
\usage{
plot_experimental_design(experimental_design)
}
\arguments{
\item{experimental_design}{an experimental design matrix rownames should be microfluidics channels, colnames should be genotypes}
}
\value{
a ggplot experimental design heatmap
}
\description{
Function to plot the experimental design
}
\examples{
\donttest{
#' plot_experimental_design()
}
}
