% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_mix_functions.R
\name{read_experimental_design}
\alias{read_experimental_design}
\title{read_experimental_design}
\usage{
read_experimental_design(experimental_design_path)
}
\arguments{
\item{experimental_design_path}{path to the experimental design .csv file}
}
\value{
a matrix of the experimental design
}
\description{
read_experimental_design
}
\examples{
\donttest{
read_experimental_design()
}
}
