% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_mix_functions.R
\name{read_SOC_locations}
\alias{read_SOC_locations}
\title{read_SOC_locations}
\usage{
read_SOC_locations(SOC_locations)
}
\arguments{
\item{SOC_locations}{path to the .csv file giving channel names as the first column and file paths to the souporcell output in the second column}
}
\value{
a matrix of SOC file paths and the channel names
}
\description{
Fucntion to read in the locations of souporcell outputs
}
\examples{
\donttest{
#' read_SOC_locations()
}
}
