% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_mix_functions.R
\name{shared_genotypes}
\alias{shared_genotypes}
\title{shared_genotypes}
\usage{
shared_genotypes(
  vcf1,
  vcf2,
  shared,
  experiment_1_name,
  experiment_2_name,
  ncounts = 10
)
}
\arguments{
\item{vcf1}{the first vcf}

\item{vcf2}{the second vcf}

\item{shared}{numeric- the number of shared genotypes between the experiments}

\item{experiment_1_name}{the name of experiment 1}

\item{experiment_2_name}{the name of experiment 2}

\item{ncounts}{numeric - the number of counts supporting each variant}
}
\value{
a data frame of the shared genotype clusters between the two experiments
}
\description{
Function that compares genotype calls across two experiments
}
\examples{
\donttest{
#' shared_genotypes()
}
}
