% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_mix_functions.R
\name{shared_genotypes_relatedness}
\alias{shared_genotypes_relatedness}
\title{this is a function that compares genotype calls across two experiments and calculates a "relatedness" score as a reimplementation of the Somalier approach Pedersen et al. 2020 (10.1186/s13073-020-00761-2)}
\usage{
shared_genotypes_relatedness(
  vcf1,
  vcf2,
  shared,
  experiment_1_name,
  experiment_2_name
)
}
\arguments{
\item{vcf1}{the first vcf}

\item{vcf2}{the second vcf}

\item{shared}{numeric- the number of shared genotypes between the experiments}

\item{experiment_1_name}{the name of experiment 1}

\item{experiment_2_name}{the name of experiment 2}
}
\value{
a data frame of the shared genotype clusters between the two experiments, returning the relatedness score.
}
\description{
this is a function that compares genotype calls across two experiments and calculates a "relatedness" score as a reimplementation of the Somalier approach Pedersen et al. 2020 (10.1186/s13073-020-00761-2)
}
\examples{
\donttest{
#' shared_genotypes_relatedness()
}
}
