% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_mix_functions.R
\name{get_shared}
\alias{get_shared}
\title{This is a function to get the number of shared genotypes between channels}
\usage{
get_shared(experimental_design)
}
\arguments{
\item{experimental_design}{design matrix rownames should be microfluidics channels, colnames should be genotypes}
}
\value{
a matrix of channels and genotypes shared
}
\description{
This is a function to get the number of shared genotypes between channels
}
\examples{
\donttest{
# get_shared()
}
}
