% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_mix_functions.R
\name{plot_cross_vaf}
\alias{plot_cross_vaf}
\title{this is a function that maps the genotype clusters to genotype names}
\usage{
plot_cross_vaf(
  experiment_1_path,
  experiment_2_path,
  experiment_1_name,
  experiment_2_name,
  ncounts = 10
)
}
\arguments{
\item{experiment_1_path}{the SOC directory for the first experiment}

\item{experiment_2_path}{the SOC directory for the second experiment}

\item{experiment_1_name}{the name of the first experiment}

\item{experiment_2_name}{the name of the second experiment}
}
\value{
matrix of cells by genotypes
}
\description{
this is a function that maps the genotype clusters to genotype names
}
\examples{
\donttest{
cells_to_genotypes()
}
}
