% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_mix_functions.R
\name{shared_genotypes}
\alias{shared_genotypes}
\title{this is a function that compares genotype calls across two experiments}
\usage{
shared_genotypes(
  experiment_1_path,
  experiment_2_path,
  shared,
  experiment_1_name,
  experiment_2_name
)
}
\arguments{
\item{experiment_1_path}{the file path to the first souporcell directory}

\item{experiment_2_path}{the file path to the second souporcell directory}

\item{shared}{numeric- the number of shared genotypes between the experiments}

\item{experiment_1_name}{the name of experiment 1}

\item{experiment_2_name}{the name of experiment 2}
}
\value{
a data frame of the shared genotype clusters between the two experiments
}
\description{
this is a function that compares genotype calls across two experiments
}
\examples{
\donttest{
#' shared_genotypes()
}
}
